PostProcessingHook added to common_algo
Please note that map-from is not an attempt to
fully support arbitrary mappings from a DDL to
Hydrolix but partial support for easing migrations
from other platforms.

Elastic needs some more review since there are
some design decisions/complications but it can
already apply transforms.

Main problem is field cardinality which is 0..* for
any Elastic field. A good interface for this would
be needed for the user to choose.

Regarding SQL, it is also an incomplete but working
implementation.