docs: tighten MD gen: options only if present
This commit changes the `HdxGroup` and `HdxCommand` objects to
suppressany output when no options table is necessary.

The `_generate_options_table` function used to return an empty string
which was always emitted. Now, instead, it returns `None` to indicate
that no options table was found, allowing the caller to skip producing
extra newlines under those circumstances.

Prior to this change, the resulting Markdown files sport the following
violation count for Markdown syntax rules:

39 MD012 no-multiple-blanks
 7 MD022 blanks-around-headings

There are no more cases of MD012 and MD022 in the resulting output now.

- https://pymarkdown.readthedocs.io/en/latest/plugins/rule_md022/
- https://pymarkdown.readthedocs.io/en/latest/plugins/rule_md012/