docs: tighten MD gen: unique headings
Generate a unique heading within each document. Use the click context to
construct a heading that identifies the resource and command context.

While this may appear to be a cosmetic fix, it enables Markdown
rendering engines to make use of HTTP link anchors for locating a user
in the document text.

- https://pymarkdown.readthedocs.io/en/latest/plugins/rule_md024/