docs: shadow: place `shadow_` in code backticks (Markdown)
The Markdown parser treats the underscore as a part of an emphasis
according to the CommonMark specification. This means that without
marking as code, the string "shadow_" is an unterminated emphasis.

In this case, there is no closing underscore (`_`) and the parser is
detecting the space adjacent to the sole underscore as some part of an
emphasis element.

This is a small cosmetic change that allows the generated Markdown to
satisfy the `pymarkdown` rules used by the Hydrolix docs system.

- https://pymarkdown.readthedocs.io/en/latest/plugins/rule_md037/