docs: migrate: no emphasis, 4 space nested lists (Markdown)
The Markdown parser (`pymdown`) in use by the documentation team expects

- a blank line between a paragraph and a list [MD032]
- no use of emphasis as the sole item in a paragraph or list [MD036]

These are small cosmetic changes that allow the generated Markdown to
satisfy the `pymarkdown` rules used by the Hydrolix docs system.

- https://pymarkdown.readthedocs.io/en/latest/plugins/rule_md032/
- https://pymarkdown.readthedocs.io/en/latest/plugins/rule_md036/

Additionally, included in this commit is a nested list indenting change
to work better with the Markdown rendering tool (Material for `mkdocs`)
used by the Hydrolix docs system.

This rendering tool expects nested unordered lists to have an
indentation level of four (4) spaces.  This deviates from the CommonMark
specification which expects two (2) spaces as expressed in MD007. The
deviation is intentional.