docs: tighten MD gen: render Options as a (sub)heading
Switch from using Markdown emphasis to using a heading, whose level is
derived from the click context depth.

The `_generate_options_table` function already accepts the click
context. In order to ensure that the Options section is placed correctly
in a hierarchy, subordinate to the calling click context, we must add a
single level to the depth.

For example, if the calling object is an H3, we must calculate the
Options section for the H3 as an H4, so that it nests correctly in the
document structure.

- https://pymarkdown.readthedocs.io/en/latest/plugins/rule_md036/