docs: tighten MD gen: drop extraneous newlines
This is the first of several commits changing the Markdown emitted from
the `click_extensions.py` code.

The Markdown parser (`pymdown`) in use by the documentation team expects
a single blank line between a paragraph and a list.

This tightens up the Markdown emission code to remove extraneous
newlines.

Prior to this change, the resulting Markdown files sport the following
violation count for Markdown syntax rules:

166 MD012 no-multiple-blanks
134 MD022 blanks-around-headings

After this change, the report looks like this:

39 MD012 no-multiple-blanks
 7 MD022 blanks-around-headings

The subsequent commit fixes remaining violations of these rules.

- https://pymarkdown.readthedocs.io/en/latest/plugins/rule_md022/
- https://pymarkdown.readthedocs.io/en/latest/plugins/rule_md012/