docs: shrink front matter; excerpt -> subtitle (de-RDMD)
The dialect of Markdown used by the ReadMe hosting (RMH) platform was
called ReadMe Markdown (RDMD) and the RMH platform supported several
keys in the front matter which we do not use in the MkDocs system.

This commit removes the unnecessary front matter keys `slug`, `category`
and `excerpt` while retaining the excerpt text as a subtitle element
available to the new MkDocs system.