docs: tighten MD gen: render Usage + Examples as headings
Switch from using Markdown emphasis to using a subheading, whose level is
derived from the click context depth.

In order to ensure that the Usage and Examples sections are placed correctly
in the document structure, add a single level to the depth calculated
handled when processing an `HdxGroup` or `HdxCommand`.

For example, if the object getting rendered to Markdown is at click
command depth 2, it will be shown as an H3. We must therefore calculate
the Usage and Examples sections as an H4, so that it nests correctly in
the document structure.

This commit fixes all `no-emphasis-as-heading` Markdown rule violations,
MD036.

- https://pymarkdown.readthedocs.io/en/latest/plugins/rule_md036/